Rearrange factories when creating new tests
