+++ apps/accounts/templates/emails/verification-request-notify.txt
@@ -1,16 +1,24 @@
 Hello
-
-Thank you for taking the time to {% if provider %}{%- comment -%} If it's an update to an existing provider. {%- endcomment -%}update the listing for {{ provider.name }} in the Green Web Dataset.{% else %}{%- comment -%} If it's a new provider request. {%- endcomment -%}complete a Green Web Dataset verification request for {{ org_name }}.{% endif %}
-
-{% if provider %}You can see your update you have submitted and its current status at the link below:{% else %}You can see your submitted request and its current status at the link below:{% endif %}
-
+Thank you for taking the time to
+{% if provider %}
+    {% comment %} If it's an update to an existing provider. {% endcomment %}update the listing for {{ provider.name }} in the Green Web Dataset.
+{% else %}
+    {% comment %} If it's a new provider request. {% endcomment %}complete a Green Web Dataset verification request for {{ org_name }}.
+{% endif %}
+{% if provider %}
+    You can see your update you have submitted and its current status at the link below:
+{% else %}
+    You can see your submitted request and its current status at the link below:
+{% endif %}
 {{ link_to_verification_request }}
-
 What happens next?
-
-We aim to review every {% if provider %}{%- comment -%} If it's an update to an existing provider. {%- endcomment -%}update{% else %}{%- comment -%} If it's a new provider request. {%- endcomment -%}new verification request{% endif %}within 3 working days. Once we have reviewed the verification request, we will contact you by email to let you know that it is approved, or that we need more information from you.
-
+We aim to review every
+{% if provider %}
+    {% comment %} If it's an update to an existing provider. {% endcomment %}update
+{% else %}
+    {% comment %} If it's a new provider request. {% endcomment %}new verification request
+{% endif %}
+within 3 working days. Once we have reviewed the verification request, we will contact you by email to let you know that it is approved, or that we need more information from you.
 If you have further questions, or need to edit your verification request,  please contact support@thegreenwebfoundation.org.
-
 Many thanks,
 Green Web Foundation
