+++ apps/accounts/tests/test_provider_request.py
@@ -585,7 +585,7 @@ def test_approve_updates_existing_provider(hosting_provider_with_sample_user):
 
 @pytest.mark.django_db
 def test_approve_updates_existing_provider_without_deleting_asns(
-    hosting_provider_with_sample_user
+    hosting_provider_with_sample_user,
 ):
     """
     Check that approving a provider request does not delete an existing ASN, but
@@ -622,7 +622,7 @@ def test_approve_updates_existing_provider_without_deleting_asns(
 
 @pytest.mark.django_db
 def test_approve_updates_existing_provider_without_deleting_ips(
-    hosting_provider_with_sample_user
+    hosting_provider_with_sample_user,
 ):
     # given: existing ips linked to our provider
     active_ip = GreenIpFactory.create(hostingprovider=hosting_provider_with_sample_user)
@@ -664,7 +664,7 @@ def test_approve_updates_existing_provider_without_deleting_ips(
 
 @pytest.mark.django_db
 def test_approve_updates_existing_provider_without_deleting_supporting_evidence(
-    hosting_provider_with_sample_user
+    hosting_provider_with_sample_user,
 ):
     # given: a provider request linked to an existing hosting provider
 
@@ -1922,4 +1922,44 @@ def test_request_from_host_provider_finishes_in_sensible_time():
     """
     hosting provider with just a country should show the city and country
     """
+    pass
+
+
+@pytest.mark.django_db
+@override_flag("provider_request", active=True)
+def test_email_sent_on_approval(
+    user,
+    client,
+    greenweb_staff_user,
+    provider_request_factory,
+    wizard_form_org_details_data,
+    wizard_form_org_location_data,
+    wizard_form_services_data,
+    wizard_form_evidence_data,
+    wizard_form_network_data,
+    wizard_form_consent,
+    wizard_form_preview,
+    mailoutbox,
+    hosting_provider_with_sample_user,
+):
+    """
+    Given: a provider request to update an existing provider
+    When: it is approved by staff, we should send an email
+    Then:
+    """
+
+    pr = provider_request_factory.create(provider=hosting_provider_with_sample_user)
+
+    client.force_login(greenweb_staff_user)
+
+    from django.urls import reverse
+
+    admin_update_path = reverse(
+        "greenweb_admin:accounts_providerrequest_change", args=[pr.id]
+    )
+
+    client.get(admin_update_path)
+
+    breakpoint()
+
     pass
