+++ conftest.py
@@ -32,81 +32,7 @@ register(gc_factories.SiteCheckFactory)
 
 register(ac_factories.SupportingEvidenceFactory)
 
-
-class ProviderRequestFactory(factory.django.DjangoModelFactory):
-    """
-    By default ProviderRequestFactory() or ProviderRequestFactory.create()
-    will return an object with no services. This is because of
-    how factory-boy manages many-to-many relationships.
-
-    Services can be set in the following ways:
-    - pr = ProviderRequestFactory(); pr.services.set(["service1", "service2"])
-    - ProviderRequestFactory.create(services=["service1", "service2"])
-    """
-
-    name = factory.Faker("word")
-    website = factory.Faker("domain_name")
-    description = factory.Faker("sentence")
-    status = ac_models.ProviderRequestStatus.OPEN
-    created_by = factory.SubFactory(gc_factories.UserFactory)
-    authorised_by_org = True
-
-    @factory.post_generation
-    def services(self, create, extracted, **kwargs):
-        """
-        This handles many-to-many relationship between ProviderRequest and Tag.
-
-        More details: https://factoryboy.readthedocs.io/en/latest/recipes.html#simple-many-to-many-relationship
-        """  # noqa
-        # nothing passed as an argument
-        if not create or not extracted:
-            return
-        # set tags
-        for tag in extracted:
-            self.services.add(tag)
-
-    class Meta:
-        model = ac_models.ProviderRequest
-
-
-class ProviderRequestLocationFactory(factory.django.DjangoModelFactory):
-    name = factory.Faker("word")
-    city = factory.Faker("city")
-    country = factory.Faker("country_code")
-    request = factory.SubFactory(ProviderRequestFactory)
-
-    class Meta:
-        model = ac_models.ProviderRequestLocation
-
-
-class ProviderRequestEvidenceFactory(factory.django.DjangoModelFactory):
-    title = factory.Faker("sentence", nb_words=3)
-    description = factory.Faker("sentence", nb_words=6)
-    link = factory.Faker("url")
-    file = None
-    type = factory.Faker("random_element", elements=ac_models.EvidenceType.values)
-    public = factory.Faker("random_element", elements=[True, False])
-    request = factory.SubFactory(ProviderRequestFactory)
-
-    class Meta:
-        model = ac_models.ProviderRequestEvidence
-
-
-class ProviderRequestIPRangeFactory(factory.django.DjangoModelFactory):
-    start = factory.Faker("ipv4")
-    end = factory.LazyAttribute(lambda o: str(ip_address(o.start) + 10))
-    request = factory.SubFactory(ProviderRequestFactory)
-
-    class Meta:
-        model = ac_models.ProviderRequestIPRange
-
-
-class ProviderRequestASNFactory(factory.django.DjangoModelFactory):
-    asn = factory.Faker("random_int")
-    request = factory.SubFactory(ProviderRequestFactory)
-
-    class Meta:
-        model = ac_models.ProviderRequestASN
+register(ac_factories.ProviderRequestFactory)
 
 
 @pytest.fixture
